[REFACTOR] organize repository folder
