+++ src/main/java/com/won/myongjiCamp/controller/api/BoardApiController.java
@@ -7,8 +7,8 @@ import com.won.myongjiCamp.dto.response.ResponseDto;
 import com.won.myongjiCamp.model.board.*;
 import com.won.myongjiCamp.model.board.role.RoleAssignment;
 import com.won.myongjiCamp.repository.*;
-import com.won.myongjiCamp.repository.board.complete.CompleteRepository;
-import com.won.myongjiCamp.repository.board.recruit.RecruitRepository;
+import com.won.myongjiCamp.repository.CompleteRepository;
+import com.won.myongjiCamp.repository.RecruitRepository;
 import com.won.myongjiCamp.service.BoardService;
 import com.won.myongjiCamp.service.CompleteService;
 import com.won.myongjiCamp.service.RecruitService;
