+++ src/main/java/com/won/myongjiCamp/repository/BoardRepository.java
@@ -1,4 +1,4 @@
-package com.won.myongjiCamp.repository.board;
+package com.won.myongjiCamp.repository;
 
 import com.won.myongjiCamp.model.board.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
