+++ src/main/java/com/won/myongjiCamp/repository/CompleteRepository.java
@@ -1,7 +1,8 @@
-package com.won.myongjiCamp.repository.board.complete;
+package com.won.myongjiCamp.repository;
 
 import com.won.myongjiCamp.model.Member;
 import com.won.myongjiCamp.model.board.CompleteBoard;
+import com.won.myongjiCamp.repository.custom.CompleteRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
