+++ src/main/java/com/won/myongjiCamp/repository/custom/CompleteRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.won.myongjiCamp.repository.board.complete;
+package com.won.myongjiCamp.repository.custom;
 
 import static com.won.myongjiCamp.model.board.QCompleteBoard.completeBoard;
