+++ src/main/java/com/won/myongjiCamp/repository/custom/RecruitRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.won.myongjiCamp.repository.board.recruit;
+package com.won.myongjiCamp.repository.custom;
 
 import static com.won.myongjiCamp.model.board.QRecruitBoard.recruitBoard;
