+++ src/main/java/com/won/myongjiCamp/service/ApplicationService.java
@@ -12,8 +12,8 @@ import com.won.myongjiCamp.model.board.RecruitStatus;
 import com.won.myongjiCamp.model.board.role.Role;
 import com.won.myongjiCamp.model.board.role.RoleAssignment;
 import com.won.myongjiCamp.repository.ApplicationRepository;
-import com.won.myongjiCamp.repository.board.BoardRepository;
-import com.won.myongjiCamp.repository.board.recruit.RecruitRepository;
+import com.won.myongjiCamp.repository.BoardRepository;
+import com.won.myongjiCamp.repository.RecruitRepository;
 import com.won.myongjiCamp.repository.RoleAssignmentRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
