+++ src/main/java/com/won/myongjiCamp/service/BoardService.java
@@ -4,8 +4,8 @@ import com.won.myongjiCamp.dto.request.BoardRequest.BoardSearchDto;
 import com.won.myongjiCamp.model.Member;
 import com.won.myongjiCamp.model.board.CompleteBoard;
 import com.won.myongjiCamp.model.board.RecruitBoard;
-import com.won.myongjiCamp.repository.board.complete.CompleteRepository;
-import com.won.myongjiCamp.repository.board.recruit.RecruitRepository;
+import com.won.myongjiCamp.repository.CompleteRepository;
+import com.won.myongjiCamp.repository.RecruitRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
