+++ src/main/java/com/won/myongjiCamp/service/CompleteService.java
@@ -5,9 +5,9 @@ import com.won.myongjiCamp.model.Member;
 import com.won.myongjiCamp.model.board.CompleteBoard;
 import com.won.myongjiCamp.model.board.Image;
 import com.won.myongjiCamp.model.board.RecruitBoard;
-import com.won.myongjiCamp.repository.board.complete.CompleteRepository;
+import com.won.myongjiCamp.repository.CompleteRepository;
 import com.won.myongjiCamp.repository.ImageRepository;
-import com.won.myongjiCamp.repository.board.recruit.RecruitRepository;
+import com.won.myongjiCamp.repository.RecruitRepository;
 import lombok.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
