+++ src/main/java/com/won/myongjiCamp/service/ScrapService.java
@@ -4,7 +4,7 @@ import com.won.myongjiCamp.dto.request.ScrapRequest;
 import com.won.myongjiCamp.model.Member;
 import com.won.myongjiCamp.model.Scrap;
 import com.won.myongjiCamp.model.board.Board;
-import com.won.myongjiCamp.repository.board.BoardRepository;
+import com.won.myongjiCamp.repository.BoardRepository;
 import com.won.myongjiCamp.repository.MemberRepository;
 import com.won.myongjiCamp.repository.ScrapRepository;
 import com.won.myongjiCamp.specification.ScrapSpecification;
